Data window fix & enhancement
The data windowing method plots the data of interest against the data
index to simplify indexing the windowed data. In the case of the descent
rate calculation method, the data is replotted against time for better
vieweing. The axes limits are then adjusted based on the limits of the
axes being used while windowing the data. Fixes have been implemented
for cases where these axes limits correspond to invalid indices, either
indices less than 1 or indices greater than the length of the data set.

Added logic to set the initial x locations of the window lines based on
the data width of the axes rather than arbitrary hard-coded locations.

Resolves #1