Rename pressure ref, finish inline documentation
Pressure reference has been changed from sea level to ground level.
Because the pressure selection workflow asks the user to identify the
ground level pressure, calling it a sea level pressure is not intuitive
and wrong in any case where we're not dropping onto the beach. That
would be nice though.

Inline documentation and explanatory comments have been added throughout